+++ Pokemon2.vcxproj
@@ -127,31 +127,31 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="BattleManager.cpp" />
-    <ClCompile Include="Game.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="Pokemon.cpp" />
+    <ClCompile Include="src\Battle\BattleManager.cpp" />
+    <ClCompile Include="src\Main\Game.cpp" />
+    <ClCompile Include="src\Character\Player\Player.cpp" />
+    <ClCompile Include="src\Pokemon\Pokemon.cpp" />
     <ClCompile Include="Pokemon2.cpp" />
-    <ClCompile Include="ProfessorOak.cpp" />
-    <ClCompile Include="Utility.cpp" />
-    <ClCompile Include="WildEncounterManager.cpp" />
+    <ClCompile Include="src\Character\ProfessorOak.cpp" />
+    <ClCompile Include="src\Utility\Utility.cpp" />
+    <ClCompile Include="src\Battle\WildEncounterManager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include=".gitattributes" />
     <None Include=".gitignore" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BattleManager.h" />
-    <ClInclude Include="BattleState.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="grass.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="Pokemon.h" />
-    <ClInclude Include="PokemonChoice.h" />
-    <ClInclude Include="PokemonType.h" />
-    <ClInclude Include="ProfessorOak.h" />
-    <ClInclude Include="Utility.h" />
-    <ClInclude Include="WildEncounterManager.h" />
+    <ClInclude Include="include\Battle\BattleManager.h" />
+    <ClInclude Include="include\Battle\BattleState.h" />
+    <ClInclude Include="include\Main\Game.h" />
+    <ClInclude Include="include\Pokemon\grass.h" />
+    <ClInclude Include="include\Character\Player\Player.h" />
+    <ClInclude Include="include\Pokemon\Pokemon.h" />
+    <ClInclude Include="include\Pokemon\PokemonChoice.h" />
+    <ClInclude Include="include\Pokemon\PokemonType.h" />
+    <ClInclude Include="include\Character\ProfessorOak.h" />
+    <ClInclude Include="include\Utility\Utility.h" />
+    <ClInclude Include="include\Battle\WildEncounterManager.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
