+++ src/Character/Player/Player.cpp
@@ -1,9 +1,9 @@
 // Player.cpp
 
-#include "Player.h"
-#include "PokemonChoice.h"
-#include "Pokemon.h"
-#include "Utility.h"
+#include "../../include/Character/Player/Player.h"
+#include "../../include/Pokemon/PokemonChoice.h"
+#include "../../include/Pokemon/Pokemon.h"
+#include "../../include/Utility/Utility.h"
 #include <iostream>
 
 using namespace std;
