+++ src/Character/ProfessorOak.cpp
@@ -1,7 +1,7 @@
 // ProfessorOak.cpp
-#include "ProfessorOak.h"
-#include "Player.h"
-#include "Utility.h"
+#include "../../include/Character/ProfessorOak.h"
+#include "../../include/Character/Player/Player.h"
+#include "../../include/Utility/Utility.h"
 #include <iostream>
 
 using namespace std;
